requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
APScheduler==3.9.1
arq==0.22
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
backports.zoneinfo==0.2.1
charset-normalizer==2.1.0
click==8.1.3
frozenlist==1.3.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
multidict==6.0.2
orjson==3.7.6
pydantic==1.9.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
sanic==22.6.0
sanic-routing==22.3.0
six==1.16.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
yarl==1.7.2
